requirements.txt
fastapi
uvicorn
pandas
python-multipart
xlrd
openpyxl
pymongo
pydantic
python-dotenv